removal smart.dm_att_tree_node table added getActiveChildren to categories and trees fixed validation bugs
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk/source@397 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ java/org.wcs.smart.query/src/org/wcs/smart/query/QueryPlugIn.java
@@ -278,9 +278,11 @@ public class QueryPlugIn extends AbstractUIPlugin {
 			@Override
 			protected IStatus run(IProgressMonitor monitor) {
 				File dir = getQueryTempDirectory();
-				File[] toDel = dir.listFiles();
-				for (int i = 0; i < toDel.length; i ++){
-					toDel[i].delete();
+				if (dir.exists() && dir.isDirectory()){
+					File[] toDel = dir.listFiles();
+					for (int i = 0; i < toDel.length; i ++){
+						toDel[i].delete();
+					}
 				}
 				return Status.OK_STATUS;
 			}
